+++ confiture-web-app/src/store/results.ts
@@ -72,6 +72,9 @@ interface ResultsStoreState {
   lastRequestFailed: boolean;
 
   lastUpdatedTopic: number;
+
+  // FIXME: refactor the way all requests are done
+  abortController: AbortController | null;
 }
 
 export const useResultsStore = defineStore("results", {
@@ -84,7 +87,8 @@ export const useResultsStore = defineStore("results", {
       currentRequestCount: 0,
       lastRequestSuccessEnd: null,
       lastRequestFailed: false,
-      lastUpdatedTopic: 1
+      lastUpdatedTopic: 1,
+      abortController: null
     };
   },
 
@@ -563,41 +567,64 @@ export const useResultsStore = defineStore("results", {
 
     async createNotCompliantItem(
       uniqueId: string,
+      // FIXME: pageId is required to update store after request
+      pageId: number,
       slug: string,
       topic: number,
       criterium: number,
       notCompliantItem: CreateNotCompliantItemData | null = null
     ): Promise<NotCompliantItem> {
-      if (notCompliantItem) {
-        return (await api
-          .post(`/api/audits/${uniqueId}/pages/${slug}/results/${topic}.${criterium}/not-compliant-items`, {
-            json: notCompliantItem
-          })
-          .json()) as NotCompliantItem;
-      }
+      const createdItem = await api
+        .post(`/api/audits/${uniqueId}/pages/${slug}/results/${topic}.${criterium}/not-compliant-items`, {
+          ...(notCompliantItem && { json: notCompliantItem })
+        })
+        .json<NotCompliantItem>();
 
-      return (await api
-        .post(`/api/audits/${uniqueId}/pages/${slug}/results/${topic}.${criterium}/not-compliant-items`)
-        .json()) as NotCompliantItem;
+      // update result in store
+      this.data?.[pageId][topic][criterium].notCompliantItems.push(createdItem);
+
+      return createdItem;
     },
 
     async updateNotCompliantItem(
       uniqueId: string,
+      // FIXME: pageId is required to update store after request
+      pageId: number,
       slug: string,
       topic: number,
       criterium: number,
       notCompliantItemId: number,
       notCompliantItem: UpdateNotCompliantItemData
-    ): Promise<NotCompliantItem> {
-      return (await api
+    ): Promise<NotCompliantItem | void> {
+      this.increaseCurrentRequestCount();
+
+      const updatedItem = await api
         .patch(`/api/audits/${uniqueId}/pages/${slug}/results/${topic}.${criterium}/not-compliant-items/${notCompliantItemId}`, {
-          json: notCompliantItem
-        })
-        .json()) as NotCompliantItem;
+          json: notCompliantItem,
+          signal: this.abortController?.signal
+        }).json<NotCompliantItem>().catch((error) => {
+          if (error.name === "AbortError") {
+            return;
+          }
+        }).finally(() => {
+          this.decreaseCurrentRequestCount();
+        });
+
+      if (updatedItem) {
+      // update result in store
+        const idx = this.data?.[pageId][topic][criterium].notCompliantItems.findIndex(item => item.id === notCompliantItemId) ?? -1;
+        if (idx !== -1) {
+          this.data?.[pageId][topic][criterium].notCompliantItems.splice(idx, 1, updatedItem);
+        }
+      }
+
+      return updatedItem;
     },
 
     async deleteNotCompliantItem(
       uniqueId: string,
+      // FIXME: pageId is required to update store after request
+      pageId: number,
       slug: string,
       topic: number,
       criterium: number,
@@ -605,6 +632,17 @@ export const useResultsStore = defineStore("results", {
     ): Promise<void> {
       await api
         .delete(`/api/audits/${uniqueId}/pages/${slug}/results/${topic}.${criterium}/not-compliant-items/${notCompliantItemId}`);
+
+      // update result in store
+      const idx = this.data?.[pageId][topic][criterium].notCompliantItems.findIndex(item => item.id === notCompliantItemId) ?? -1;
+      if (idx !== -1) {
+        this.data?.[pageId][topic][criterium].notCompliantItems.splice(idx, 1);
+      }
+    },
+
+    abortRequest() {
+      this.abortController?.abort();
+      this.abortController = new AbortController();
     }
   }
 });