+++ CHANGELOG.md
@@ -4,6 +4,25 @@
 
 <h2 class="fr-sr-only" id="2026">2026</h2>
 
+### 10/08/2026
+
+#### <span aria-hidden="true">⚙️</span> Autres changements
+
+- Ajout d’une page d’aide pour expliquer comment utiliser l’onglet «&nbsp;Éléments transverses&nbsp;» ([#1561](https://github.com/DISIC/Ara/pull/1561))
+
+### 29/07/2026
+
+#### <span aria-hidden="true">⚙️</span> Autres changements
+
+- Ajouter les actions liées à un audit dans la page "Tableau de bord d'un audit  ([#1557](https://github.com/DISIC/Ara/pull/1557))
+
+### 24/07/2026
+
+#### <span aria-hidden="true">⚙️</span> Autres changements
+
+- Ajouter le placeholder "Entité ayant réalisé l'audit" dans la déclaration ([#1555](https://github.com/DISIC/Ara/pull/1555))
+- Les liens du haut du pied de page s’affichent quelque soit l’état de connexion de l’usager (mais pas dans le rapport) ([#1570](https://github.com/DISIC/Ara/pull/1570))
+
 ### 22/07/2026
 
 #### <span aria-hidden="true">🚀</span> Nouvelles fonctionnalités